Add single dim matrix tests for LC73
Hi, 

First of all thank you a lot for your online course, I've completed it recently and I must say it's very instructive and entertaining at the same time.

However I've noticed for some specific problems, even though we pass the test cases on local machine, it's not enough for leetcode. Most of the time it's just some small mistakes like variable names, for-loop limits.. and sometimes it's a bit more complex (matrix row & column indexes for example), it can take us a lot of time to debug.

A solution is to add more test cases (starting with very obvious edge cases). 
I can try to open PRs for tests on other problems which helped me to debug, but if it doesn't match your philosophy and you want to keep the same tests as leetcode's ones, then I can understand and let's just close this PR.

Thank you again and don't hesitate to do more courses on new problems, it really helps us a lot (well at least for me).